ide: Improve line length settings
Currently "mandatory wrapping" is configured in the project settings, which is not what everyone wants.

Example: It took me a surprisingly long amount of time to understand a line of code, because it looked like two lines. I've not used wrapping in the last 10 years... or ever.

At the same time let's add a python ruler for 80 characters, since it's the de-facto-standard. Not only de-factory, it is the standard. See PEP-8.

Signed-off-by: Alexander Lanin <Alexander.Lanin@etas.com>
Issue-ref: see #76