sphinx: Added sphinx_extension folders
The sphinx_extension folders are moved over with this commit.
This adds multiple warnings for different requirements.
These warnings check if certain parameters are fulfilled when the
documentation builds.

Note: The warnings are NOW not interpreted as errors, to allow for
a successfull build but give stakeholders the opportunity to see
what checks for what reasons fail.