High level Architecture for LoLa (#374)
* communication: High level architecture

Draft of the HLA

* communication: Remove req links

Draft of the HLA

* communication: bugfix build

- create interface
- resolve links

* ipc: Move architecture to own directory

An own directory helps better structure
the toc and sort assets.

* ipc: Fix the TOC

Do not show all headings in feature request in TOC.
Show architecture and roadmap in TOC.

* communication: Separate architecture

Separate architecture into feature and component.
For now component architecture is stored in feature tree because
of tooling issues.

---------

Co-authored-by: Jochen Hoenle <jochen.hoenle@bmw.de>